+++ html/employee/administrative/informationList.html
@@ -27,8 +27,8 @@
 </div>
 
 <script type="module">
-  import { renderGreeting } from '../../../js/utils/uiUtils/messageManager/employee/administrative/administrative.js';
-  import { renderTable } from '../../../js/utils/uiUtils/tableRenderer/employee/administrative/administrative.js';
+  import { renderGreeting } from '../../../js/utils/uiUtils/messageManager/employee/administrative/message.js';
+  import { renderTable } from '../../../js/utils/uiUtils/tableRenderer/employee/administrative/informationList.js';
 
   document.addEventListener("DOMContentLoaded", () => {
     renderGreeting();
