+++ html/employee/customerInformationManagement/customerInformationManagement.html
@@ -15,8 +15,8 @@
 </div>
 
 <script type="module">
-  import { renderGreeting } from '../../../js/utils/uiUtils/messageManager/employee/customerInformationManagement/customerInformationManagement.js';
-  import { renderButtons } from '../../../js/utils/uiUtils/buttonManager/employee/customerInformationManagement/customerInformationManagement.js';
+  import { renderGreeting } from '../../../js/utils/uiUtils/messageManager/employee/customerInformationManagement/message.js';
+  import { renderButtons } from '../../../js/utils/uiUtils/buttonManager/employee/customerInformationManagement/home.js';
 
   // Greeting 메시지와 버튼 목록을 로드하여 표시
   document.addEventListener("DOMContentLoaded", () => {
