+++ html/employee/customerInformationManagement/informationList.html
@@ -27,8 +27,8 @@
 </div>
 
 <script type="module">
-  import { renderGreeting } from '../../../js/utils/uiUtils/messageManager/employee/customerInformationManagement/customerInformationManagement.js';
-  import { renderTable } from '../../../js/utils/uiUtils/tableRenderer/employee/customerInformationManagement/customerInformationManagement.js';
+  import { renderGreeting } from '../../../js/utils/uiUtils/messageManager/employee/customerInformationManagement/message.js';
+  import { renderTable } from '../../../js/utils/uiUtils/tableRenderer/employee/customerInformationManagement/informationList.js';
 
   document.addEventListener("DOMContentLoaded", () => {
     renderGreeting();
