+++ html/employee/managementPlanning/informationList.html
@@ -27,8 +27,8 @@
 </div>
 
 <script type="module">
-  import { renderGreeting } from '../../../js/utils/uiUtils/messageManager/employee/managementPlanning/managementPlanning.js';
-  import { renderTable } from '../../../js/utils/uiUtils/tableRenderer/employee/managementPlanning/managementPlanning.js';
+  import { renderGreeting } from '../../../js/utils/uiUtils/messageManager/employee/managementPlanning/message.js';
+  import { renderTable } from '../../../js/utils/uiUtils/tableRenderer/employee/managementPlanning/informationList.js';
 
   document.addEventListener("DOMContentLoaded", () => {
     renderGreeting();
