+++ html/employee/managementPlanning/managementPlanning.html
@@ -15,8 +15,8 @@
 </div>
 
 <script type="module">
-  import { renderGreeting } from '../../../js/utils/uiUtils/messageManager/employee/managementPlanning/managementPlanning.js';
-  import { renderButtons } from '../../../js/utils/uiUtils/buttonManager/employee/managementPlanning/managementPlanning.js';
+  import { renderGreeting } from '../../../js/utils/uiUtils/messageManager/employee/managementPlanning/message.js';
+  import { renderButtons } from '../../../js/utils/uiUtils/buttonManager/employee/managementPlanning/home.js';
 
   // Greeting 메시지와 버튼 목록을 로드하여 표시
   document.addEventListener("DOMContentLoaded", () => {
