+++ html/partnerCompany/informationList.html
@@ -27,8 +27,8 @@
 </div>
 
 <script type="module">
-  import { renderGreeting } from '../../js/utils/uiUtils/messageManager/partnerCompany/partnerCompany.js';
-  import { renderTable } from '../../js/utils/uiUtils/tableRenderer/partnerCompany/partnerCompany.js';
+  import { renderGreeting } from '../../js/utils/uiUtils/messageManager/partnerCompany/message.js';
+  import { renderTable } from '../../js/utils/uiUtils/tableRenderer/partnerCompany/informationList.js';
 
   document.addEventListener("DOMContentLoaded", () => {
     renderGreeting();
