+++ html/partnerCompany/partnerCompany.html
@@ -15,8 +15,8 @@
 </div>
 
 <script type="module">
-  import { renderGreeting } from '../../js/utils/uiUtils/messageManager/partnerCompany/partnerCompany.js';
-  import { renderButtons } from '../../js/utils/uiUtils/buttonManager/partnerCompany/partnerCompany.js';
+  import { renderGreeting } from '../../js/utils/uiUtils/messageManager/partnerCompany/message.js';
+  import { renderButtons } from '../../js/utils/uiUtils/buttonManager/partnerCompany/home.js';
 
   // Greeting 메시지와 버튼 목록을 로드하여 표시
   document.addEventListener("DOMContentLoaded", () => {
