+++ js/utils/uiUtils/buttonManager/employee/administrative/home.js
@@ -1,19 +1,19 @@
-import { BUTTON } from '../../../../../../config/employee/managementPlanning/managementPlanning.js';
-import { viewDepartmentListAll } from '../../../tableRenderer/employee/managementPlanning/managementPlanning.js';
+import { BUTTON } from '../../../../../../config/employee/administrative/administrative.js';
+import { viewOfficeSupplyListAll } from '../../../tableRenderer/employee/administrative/informationList.js';
 
 export const renderButtons = () => {
   // BUTTON 정의에 따라 ADMINISTRATIVE.HOME 버튼들을 렌더링
-  initialButtons(BUTTON.TASK.EMPLOYEE.MANAGEMENTPLANNING.HOME, managementPlanningTaskMapper);
+  initialButtons(BUTTON.TASK.EMPLOYEE.ADMINISTRATIVE.HOME, administrativeTaskMapper);
 };
 
 // 버튼 동작 정의
-const viewDepartment = async () => {
-  await viewDepartmentListAll(); // 데이터 가져오기 및 sessionStorage 저장
+const viewOfficeSupply = async () => {
+  await viewOfficeSupplyListAll(); // 데이터 가져오기 및 sessionStorage 저장
 };
 
 // 버튼 동작 매핑 객체
-const managementPlanningTaskMapper = {
-  DEPARTMENT_LIST: viewDepartment // '집기 비품 재고 관리' 버튼과 함수 연결
+const administrativeTaskMapper = {
+  OFFICESUPPLY_LIST: viewOfficeSupply // '집기 비품 재고 관리' 버튼과 함수 연결
 };
 
 // 버튼 생성 및 렌더링 함수
