+++ js/utils/uiUtils/messageManager/partnerCompany/message.js
