+++ js/utils/uiUtils/tableRenderer/partnerCompany/informationList.js
